+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 * nostr: add `Nip05Profile` struct ([Yuki Kishimoto])
 * nostr: add `nip05::profile` function ([Yuki Kishimoto])
 * nostr: add `LEN` const to `PublicKey`, `SecretKey` and `EncryptedSecretKey` ([Yuki Kishimoto])
+* nostr: add `Profile<T>` struct ([Yuki Kishimoto])
 * pool: add `Output<T>` struct ([Yuki Kishimoto])
 * pool: add `Output<EventId>::id` and `Output<SubscriptionId>::id` methods ([Yuki Kishimoto])
 * signer: add `NostrSigner::unwrap_gift_wrap` method ([Yuki Kishimoto])